Agents/PDF/pyproject.toml
[project]
name = "pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiokafka>=0.12.0",
    "chromadb>=1.0.20",
    "fastapi>=0.116.1",
    "graphviz>=0.21",
    "kafka>=1.3.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "youtube-transcript-api>=1.2.2",
]
